Miscellanous changes
add repl to task_handler

Possibility to configure log working directory
