+++ acting/Cargo.toml
@@ -13,8 +13,9 @@ tokio = { version = "1.5.0", features = ["full", "time"] }
 tokio-stream = "0.1.6"
 async-recursion = "0.3.2"
 im = "15.0.0"
-log =  { version = "0.4", features = ["std"] }
 lazy_static = "1.4.0"
+chrono = "0.4.19"
+log =  { version = "0.4", features = ["std"] }
