+++ apps/src/bin/unnamed_supervisor.rs
@@ -16,11 +16,11 @@ use ompas_lisp::core::*;
 use ompas_lisp::modules::_type::CtxType;
 use ompas_lisp::modules::counter::CtxCounter;
 use ompas_lisp::modules::doc::{CtxDoc, Documentation};
-use ompas_lisp::modules::io::repl::{spawn_log, spawn_repl};
 use ompas_lisp::modules::io::CtxIo;
 use ompas_lisp::modules::math::CtxMath;
 use ompas_lisp::modules::utils::CtxUtils;
-use ompas_utils::{log, task_handler};
+use ompas_lisp::repl::{spawn_log, spawn_repl};
+use ompas_utils::task_handler;
 
 pub const TOKIO_CHANNEL_SIZE: usize = 65_384;
 
@@ -55,7 +55,6 @@ async fn main() {
     if opt.debug {
         activate_debug();
     }
-    log::init().expect("Error while initiating logger.");
     //test_lib_model(&opt);
     lisp_interpreter(opt.log).await;
 }
@@ -74,7 +73,9 @@ pub async fn lisp_interpreter(log: Option<PathBuf>) {
         .await
         .expect("error while spawning repl");
 
-    let sender_log = spawn_log().await.expect("error while spawning log task");
+    let sender_log = spawn_log(log.clone())
+        .await
+        .expect("error while spawning log task");
 
     let (mut root_env, mut ctxs, mut lisp_init) = LEnv::root();
     let mut ctx_doc = CtxDoc::default();
@@ -84,7 +85,7 @@ pub async fn lisp_interpreter(log: Option<PathBuf>) {
     let ctx_counter = CtxCounter::default();
     let _ctx_godot = CtxGodot::default();
     let ctx_utils = CtxUtils::default();
-    let (ctx_rae, ctx_rae_monitor) = init_ctx_rae(Box::new(PlatformGodot::default()));
+    let (ctx_rae, ctx_rae_monitor) = init_ctx_rae(Box::new(PlatformGodot::default()), log.clone());
     //Insert the doc for the different contexts.
     ctx_doc.insert_doc(CtxIo::documentation());
     ctx_doc.insert_doc(CtxMath::documentation());
@@ -107,6 +108,7 @@ pub async fn lisp_interpreter(log: Option<PathBuf>) {
         agenda: Default::default(),
     };
 
+    load_module(&mut root_env, &mut ctxs, ctx_utils, &mut lisp_init);
     load_module(&mut root_env, &mut ctxs, ctx_doc, &mut lisp_init);
     load_module(&mut root_env, &mut ctxs, ctx_io, &mut lisp_init);
     load_module(&mut root_env, &mut ctxs, ctx_math, &mut lisp_init);
@@ -114,7 +116,6 @@ pub async fn lisp_interpreter(log: Option<PathBuf>) {
     load_module(&mut root_env, &mut ctxs, ctx_counter, &mut lisp_init);
     load_module(&mut root_env, &mut ctxs, ctx_rae, &mut lisp_init);
     load_module(&mut root_env, &mut ctxs, ctx_rae_monitor, &mut lisp_init);
-    load_module(&mut root_env, &mut ctxs, ctx_utils, &mut lisp_init);
 
     let env = &mut root_env.clone();
     //println!("{}", lisp_init.begin_lisp());
