+++ lisp/src/language.rs
@@ -97,171 +97,6 @@ pub mod scheme_primitives {
 
 //MACRO
 ///Problem during expansion
-pub mod scheme_macro {
-    pub const MACRO_LET: &str = "(defmacro let \
-                                           (lambda (bindings body) \
-                                                   (begin (define unzipped (unzip bindings)) \
-                                                          (define keys (car unzipped)) \
-                                                          (define values (cadr unzipped)) \
-                                                          (cons (quasiquote (lambda (unquote keys) \
-                                                                                    (unquote body))) values))))";
-
-    pub const MACRO_LET_STAR: &str = "(defmacro let* \
-                                                (lambda (bindings body) \
-                                                        (if (= (length bindings) 1) \
-                                                            (cons (quasiquote (lambda (unquote (list (caar bindings))) \
-                                                                                      (unquote body))) \
-                                                                  (cdar bindings)) \
-                                                            (cons (quasiquote (lambda (unquote (list (caar bindings))) \
-                                                                                      (let* (unquote (cdr bindings)) \
-                                                                                            (unquote body)))) \
-                                                                  (cdar bindings)))))";
-
-    //TODO: find a way to make it a lambda
-    pub const MACRO_APPLY: &str = "(defmacro apply\
-                                              (lambda (f args)\
-                                                      (cons f args)))";
-
-    pub const MACRO_AND: &str = "(defmacro and (lambda args\
-                                                (if (null? args)\
-                                                    nil\
-                                                    (if (= (length args) 1)\
-                                                        (car args)\
-                                                        (quasiquote (if (unquote (car args))\
-                                                                        (and (unquote (cdr args)))\
-                                                                        nil))))))";
-
-    pub const MACRO_OR: &str= "(defmacro or (lambda args\
-                                                (if (null? args)\
-                                                    nil\
-                                                    (if (= (length args) 1)\
-                                                        (car args)\
-                                                        (quasiquote (if (unquote (car args))\
-                                                                        true \
-                                                                        (or (unquote (cdr args)))))))))";
-
-    pub const MACRO_NEQ: &str = "(defmacro neq (lambda (a b)\
-                                            (! (= a b))))";
-
-    pub const MACRO_NEQ_SHORT: &str = "(defmacro != (lambda (a b) \
-                                                (neq a b )))";
-
-    pub const MACRO_COND: &str = "(defmacro cond (lambda exprs \
-                                                (if (null? exprs) \
-                                                    nil \
-                                                    (if (= (caar exprs) (quote else)) \
-                                                        (cadar exprs) \
-                                                        (quasiquote \
-                                                            (if (unquote (caar exprs)) \
-                                                                (unquote (cadar exprs)) \
-                                                                (cond (unquote (cdr exprs)))))))))";
-
-    pub const MACRO_AWAIT_ASYNC: &str = "(defmacro await-async (lambda (x) `(await (async ,x))))";
-
-    pub const MACRO_FOR: &str = "(defmacro for (lambda args \
-    (let ((_i_ (get-list args 0)) \
-            (_list_ (get-list args 2)) \
-            (_body_ (get-list args 3))) \
-        `(let ((_f_loop_ (lambda args \
-            (if (null? args) \
-                nil \
-                (let ((,_i_ (car args))) \
-                    (begin \
-                        ,_body_ \
-                        (_f_loop_ (cdr args)))))))) \
-            (_f_loop_ ,_list_)))))";
-}
-
-pub mod scheme_lambda {
-    pub const LAMBDA_AND: &str = "(define and (lambda args \
-                                                   (if (null? args) true \
-                                                       (if (= (length args) 1) (car args)
-                                                           (if (car args) (and (cdr args)) nil)))))";
-    pub const LAMBDA_OR: &str = "(define or (lambda args \
-                                                   (if (null? args) true \
-                                                       (if (= (length args) 1) (car args)
-                                                           (if (car args) true (or (cdr args)))))))";
-
-    pub const LAMBDA_COND: &str = "(define cond (lambda x \
-                                            (if (null? x)\
-                                                nil\
-                                                (let ((a (car x))\
-                                                       (tests (car a))\
-                                                       (expr (cdr a)))\
-                                                      (if tests expr (cond (cdr x))))))";
-
-    pub const LAMBDA_COMBINE:  &str = "(define combine (lambda (f)
-                                                                (lambda (x y)
-                                                                        (if (null? x) (quote ())
-                                                                            (f (list (car x) (car y))
-                                                                            ((combine f) (cdr x) (cdr y)))))))";
-    pub const LAMBDA_ZIP: &str = " (define zip (lambda (l1 l2)\
-                                                        (if (or (null? l1)\
-                                                                (null? l2))\
-                                                             nil\
-                                                             (cons (list (car l1)\
-                                                                         (car l2))\
-                                                                   (zip (cdr l1)\
-                                                                        (cdr l2)))))))";
-
-    /*pub const LAMBDA_CAAR: &str = "(define caar (lambda (l1) (car (car l1))))";
-
-    pub const LAMBDA_CADR: &str = "(define cadr (lambda (l1) (car (cdr l1))))";
-
-    pub const LAMBDA_CDAR: &str = "(define cdar (lambda (l1) (cdr (car l1))))";
-
-    pub const LAMBDA_CDDR: &str = "(define cddr (lambda (l1) (cdr (cdr l1))))";
-
-    pub const LAMBDA_CADAR: &str = "(define cadar (lambda (l1) (car (cdr (car l1)))))";
-
-    pub const LAMBDA_CADADR: &str = "(define cadadr (lambda (l1) (car (cdr (car (cdr l1))))))";
-
-    pub const LAMBDA_CDADR: &str = "(define cdadr (lambda (l1) (cdr (car (cdr l1)))))";
-
-    pub const LAMBDA_CADADDR: &str =
-        "(define cadaddr (lambda (l1) (car (cdr (car (cdr (cdr l1)))))))";*/
-
-    pub const MACRO_CAAR: &str =
-        "(defmacro caar (lambda (x) (quasiquote (car (car (unquote x))))))";
-    pub const MACRO_CADR: &str =
-        "(defmacro cadr (lambda (x) (quasiquote (car (cdr (unquote x))))))";
-    pub const MACRO_CDAR: &str =
-        "(defmacro cdar (lambda (x) (quasiquote (cdr (car (unquote x))))))";
-    pub const MACRO_CDDR: &str =
-        "(defmacro cddr (lambda (x) (quasiquote (cdr (cdr (unquote x))))))";
-    pub const MACRO_CAADR: &str =
-        "(defmacro caadr (lambda (x) (quasiquote (car (car (cdr(unquote x)))))))";
-    pub const MACRO_CADDR: &str =
-        "(defmacro caddr (lambda (x) (quasiquote (car (cdr (cdr (unquote x)))))))";
-    pub const MACRO_CADAR: &str =
-        "(defmacro cadar (lambda (x) (quasiquote (car (cdr (car (unquote x)))))))";
-    pub const MACRO_CADADR: &str =
-        "(defmacro cadadr (lambda (x) (quasiquote (car (cdr (car (cdr (unquote x))))))))";
-    pub const MACRO_CDADR: &str =
-        "(defmacro cdadr (lambda (x) (quasiquote (cdr (car (cdr (unquote x)))))))";
-    pub const MACRO_CADADDR: &str =
-        "(defmacro cadaddr (lambda (x) (quasiquote (car (cdr (car (cdr (cdr (unquote x)))))))))";
-
-    pub const LAMBDA_UNZIP: &str = "(define unzip
-        (lambda lists
-                (begin
-                    (define firsts (lambda lists
-                                            (if (null? lists)
-                                                nil
-                                                (cons (caar lists)
-                                                      (firsts (cdr lists))))))
-                    (define seconds (lambda lists
-                                            (if (null? lists)
-                                            nil
-                                            (cons (cadar lists)
-                                                  (seconds (cdr lists))))))
-                    (list (firsts lists) (seconds lists)))))";
-
-    pub const LAMBDA_MAPF : &str = "(define mapf (lambda (f seq)\
-                                                         (if (null? seq)\
-                                                         nil\
-                                                         (cons (eval (cons f (car seq))) (mapf f (cdr seq))))))";
-}
 //Documentation
 pub mod doc {
     //TODO: Write the doc for the basic functions
