+++ lisp/src/lib.rs
@@ -5,4 +5,7 @@ pub mod core;
 pub mod functions;
 pub mod language;
 pub mod modules;
+pub mod repl;
 pub mod structs;
+
+const TOKIO_CHANNEL_SIZE: usize = 16_384;
