+++ lisp/src/modules/io.rs
@@ -16,8 +16,6 @@ LANGUAGE
 
 //TODO: [mod] add the possibility to redirect log output to a file or any type of input
 
-const TOKIO_CHANNEL_SIZE: usize = 16_384;
-
 const MOD_IO: &str = "mod-io";
 const DOC_MOD_IO: &str = "Module than handles input/output functions.";
 const DOC_MOD_IO_VERBOSE: &str = "functions:\n\
@@ -203,250 +201,6 @@ impl Documentation for CtxIo {
             LHelp::new_verbose(READ, DOC_READ, DOC_READ_VERBOSE),
             LHelp::new_verbose(WRITE, DOC_WRITE, DOC_WRITE_VERBOSE),
         ]
-    }
-}
-
-pub mod repl {
-    //! Contains the function to handle the REPL of the project.
-    //! The repl is based on the project rustyline.
-    //!
-    //! It contains only one function (for the moment): run that takes two arguments.
-    use crate::language::scheme_primitives::NIL;
-    use crate::modules::io::TOKIO_CHANNEL_SIZE;
-    use chrono::{DateTime, Utc};
-    use ompas_utils::task_handler::subscribe_new_task;
-    use rustyline::error::ReadlineError;
-    use rustyline::Editor;
-    use std::fs::OpenOptions;
-    use std::io::Write;
-    use std::{env, fs};
-    use tokio::sync::mpsc::{self, Receiver, Sender};
-
-    ///Spawn repl task
-    pub async fn spawn_repl(sender: Sender<String>) -> Option<Sender<String>> {
-        let (sender_repl, receiver_repl) = mpsc::channel(TOKIO_CHANNEL_SIZE);
-        tokio::spawn(async move {
-            repl(sender, receiver_repl).await;
-        });
-
-        Some(sender_repl)
-    }
-
-    /// Spawn stdin task.
-    /// Not used anymore.
-    pub async fn spawn_stdin(sender: Sender<String>) -> Option<Sender<String>> {
-        let (sender_stdin, receiver_stdin) = mpsc::channel(TOKIO_CHANNEL_SIZE);
-        tokio::spawn(async move {
-            stdin(sender, receiver_stdin).await;
-        });
-
-        Some(sender_stdin)
-    }
-
-    /// Spawn stdout task
-    /// Not used anymore.
-    pub async fn spawn_stdout() -> Option<Sender<String>> {
-        let (sender_stdout, receiver_stdout): (Sender<String>, Receiver<String>) =
-            mpsc::channel(TOKIO_CHANNEL_SIZE);
-
-        tokio::spawn(async move {
-            output(receiver_stdout).await;
-        });
-
-        Some(sender_stdout)
-    }
-
-    /// Spawn the log task
-    pub async fn spawn_log() -> Option<Sender<String>> {
-        let (sender_log, receiver_log) = mpsc::channel(TOKIO_CHANNEL_SIZE);
-
-        tokio::spawn(async move {
-            log(receiver_log).await;
-        });
-
-        Some(sender_log)
-    }
-
-    /// Function to handle the repl.
-    /// ### functioning:
-    /// loop waiting for an object on *stdin*
-    /// ### args
-    /// - sender: channel object to send string to lisp interpreter.
-    /// - receiver: channel object to receive ack from lisp interpreter after evaluation.
-    /// Used for synchronization.
-    async fn stdin(sender: Sender<String>, mut receiver: Receiver<String>) {
-        let mut rl = Editor::<()>::new();
-        if rl.load_history("history.txt").is_err() {
-            println!("No previous history.");
-        }
-
-        loop {
-            let readline = rl.readline(">> ");
-
-            match readline {
-                Ok(string) => {
-                    rl.add_history_entry(string.clone());
-                    sender
-                        .send(format!("repl:{}", string))
-                        .await
-                        .expect("couldn't send lisp command");
-                    let buffer = receiver.recv().await.expect("error receiving");
-                    assert_eq!(
-                        buffer, "ACK",
-                        "should receive an ack from Lisp Intrepretor and nothing else"
-                    );
-                    //println!("repl ack: {}", buffer);
-                }
-                Err(ReadlineError::Interrupted) => {
-                    println!("CTRL-C");
-                    break;
-                }
-                Err(ReadlineError::Eof) => {
-                    println!("CTRL-D");
-                    break;
-                }
-                Err(err) => {
-                    println!("Error: {:?}", err);
-                    break;
-                }
-            }
-        }
-        sender
-            .send("exit".to_string())
-            .await
-            .expect("couldn't send exit msg");
-        rl.save_history("history.txt").unwrap();
-    }
-
-    async fn log(mut receiver: Receiver<String>) {
-        let date: DateTime<Utc> = Utc::now() + chrono::Duration::hours(2);
-        let string_date = date.format("%Y-%m-%d_%H-%M-%S").to_string();
-        let home_dir = match env::var("HOME") {
-            Ok(val) => val,
-            Err(_) => ".".to_string(),
-        };
-        let dir_path = format!("{}/ompas/lisp_logs", home_dir);
-
-        fs::create_dir_all(&dir_path).expect("could not create logs directory");
-        let mut file = OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .open(format!("{}/log_{}", &dir_path, string_date))
-            .expect("error creating log file");
-
-        let mut end_receiver = subscribe_new_task();
-
-        loop {
-            tokio::select! {
-                buffer = receiver.recv() => {
-                    let buffer = match buffer {
-                        None => {
-                            eprintln!("log task stopped working");
-                            break;
-                        }
-                        Some(b) => b,
-                    };
-                file.write_all(format!("{}\n", buffer).as_bytes())
-                    .expect("could not write to log file");
-                }
-                _ = end_receiver.recv() => {
-                    println!("log task ended");
-                    break;
-                }
-            }
-        }
-    }
-
-    /// Function to handle the repl.
-    /// ### functioning:
-    /// loop waiting for an object on *stdin*
-    /// ### args
-    /// - sender: channel object to send string to lisp interpreter.
-    /// - receiver: channel object to receive ack from lisp interpreter after evaluation.
-    /// Used for synchronization.
-    async fn repl(sender: Sender<String>, mut receiver: Receiver<String>) {
-        let mut rl = Editor::<()>::new();
-        if rl.load_history("history.txt").is_err() {
-            println!("No previous history.");
-        }
-
-        loop {
-            //TODO: to handle side effects when other user wants to connect to lisp.
-            //TODO: add multiline support
-            let readline = rl.readline(">> ");
-
-            match readline {
-                Ok(string) => {
-                    rl.add_history_entry(string.clone());
-                    sender
-                        .send(format!("repl:{}", string))
-                        .await
-                        .expect("couldn't send lisp command");
-                    let buffer = match receiver.recv().await {
-                        None => {
-                            eprintln!("repl task stopped working");
-                            break;
-                        }
-                        Some(b) => b,
-                    };
-                    if buffer != NIL {
-                        println!("LI>> {}", buffer);
-                    }
-
-                    /*assert_eq!(
-                        buffer, "ACK",
-                        "should receive an ack from Lisp Intrepretor and nothing else"
-                    );*/
-                    //println!("repl ack: {}", buffer);
-                }
-                Err(ReadlineError::Interrupted) => {
-                    println!("CTRL-C");
-                    break;
-                }
-                Err(ReadlineError::Eof) => {
-                    println!("CTRL-D");
-                    break;
-                }
-                Err(err) => {
-                    println!("Error: {:?}", err);
-                    break;
-                }
-            }
-        }
-        sender
-            .send("exit".to_string())
-            .await
-            .expect("couldn't send exit msg");
-        rl.save_history("history.txt").unwrap();
-    }
-
-    pub const EXIT_CODE_STDOUT: &str = "EXIT";
-
-    async fn output(mut receiver: Receiver<String>) {
-        let mut end_receiver = ompas_utils::task_handler::subscribe_new_task();
-        println!("output launched");
-        loop {
-            tokio::select! {
-                str = receiver.recv() => {
-                    let str = str.expect("could not receive");
-
-                    let mut stdout = std::io::stdout();
-                    stdout.lock();
-                    stdout
-                        .write_all(format!("{}\n", str).as_bytes())
-                        .expect("could not print to stdout");
-                    drop(stdout);
-                }
-                _ = end_receiver.recv() => {
-                    println!("output task ended");
-                    break;
-                }
-            }
-
-            //TODO: check if it always works.
-            //print!("{}\n", str);
-        }
     }
 }
